fix build for FreeBSD

+++ test/mtdev-test.c
@@ -27,12 +27,13 @@
  ****************************************************************************/
 
 #include <mtdev.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <unistd.h>
 #include <fcntl.h>
 
 /* year-proof millisecond event time */
-typedef __u64 mstime_t;
+typedef uint64_t mstime_t;
 
 static int use_event(const struct input_event *ev)
 {
@@ -51,7 +52,7 @@ static void print_event(const struct input_event *ev)
 	if (ev->type == EV_ABS && ev->code == ABS_MT_SLOT)
 		slot = ev->value;
 	fprintf(stderr, "%012llx %02d %01d %04x %d\n",
-		evtime, slot, ev->type, ev->code, ev->value);
+		(unsigned long long)evtime, slot, ev->type, ev->code, ev->value);
 }
 
 #define CHECK(dev, name)			\
